# Optimize lookup performance by avoiding String(describing:_)