Introduces `struct Path` as an internal abstraction for key paths, to replace `[String]`. This is required to keep the key paths `Hashable` without using `String(describing:)`.

Basic lookup benchmarks (I used the `testTomlExample` test 10000 times in a loop, excluding parsing), indicate the following speed increase:
- in debug mode: ~27s → ~20s = 25% less time
- in release mode: ~23s → ~14s = 40% less time